rename file to plural
