+++ protocol-designer/src/atoms/index.ts
@@ -1,1 +1,1 @@
-export * from './constant'
+export * from './constants'
